test: relocate stress suite under tests/agent_server/
