+++ tests/agent_server/stress/test_conversation_listing.py
@@ -25,9 +25,9 @@ from openhands.agent_server.conversation_service import ConversationService
 from openhands.agent_server.models import StartConversationRequest
 from openhands.sdk import Agent
 from openhands.sdk.workspace import LocalWorkspace
-from tests.stress.budgets import CONVERSATION_LISTING
-from tests.stress.probe import ResourceProbe
-from tests.stress.scripts import placeholder_llm
+from tests.agent_server.stress.budgets import CONVERSATION_LISTING
+from tests.agent_server.stress.probe import ResourceProbe
+from tests.agent_server.stress.scripts import placeholder_llm
 
 
 pytestmark = pytest.mark.stress
